2 : Edit/Delete issues

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,22 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
+    	User user = (User) session.getAttribute("loggeduser");
+    	
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        if(image.getUser().getId() != user.getId()) {
+        	String editError = "Only the owner of the image can edit the image";
+        	model.addAttribute("editError",editError);
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+        	return "images/image";
+        }else {
+        	String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -132,7 +141,7 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/" + updatedImage.getId();
     }
 
 
@@ -140,9 +149,19 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,HttpSession session,Model model) {
+    	User user = (User) session.getAttribute("loggeduser");
+    	Image image = imageService.getImage(imageId);
+    	if(image.getUser().getId() == user.getId()) {
+    		imageService.deleteImage(imageId);
+    		return "redirect:/images";
+    	}else {
+    		String deleteError = "Only the owner of the image can delete the image";
+    		model.addAttribute("deleteError",deleteError);
+    		model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+        	return "images/image";
+    	}
     }
